# openapi/tktsync.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TktSync API",
    "version": "1.0.0",
    "description": "Authoritative TktSync API contract for implemented M3, M4, and M5 Admin and Partner surfaces."
  },
  "paths": {
...
    "/api/v1/selection/session": {
      "get": {
        "operationId": "selectionGetSession",
        "security": [
          {
            "SelectionBearer": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response.",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControlNoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectionSession"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/selection/event": {
      "get": {
        "operationId": "selectionGetEvent",
        "security": [
          {
            "SelectionBearer": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response.",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControlNoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartnerEvent"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/selection/layout": {
      "get": {
        "operationId": "selectionGetLayout",
        "security": [
          {
            "SelectionBearer": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response.",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControlNoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericObject"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/selection/availability": {
      "get": {
        "operationId": "selectionGetAvailability",
        "security": [
          {
            "SelectionBearer": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response.",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControlNoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartnerAvailability"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/selection/reservations": {
      "post": {
        "operationId": "selectionCreateReservation",
        "security": [
          {
            "SelectionBearer": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/XRequestID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectionReservationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Authoritative buyer-scoped hold created.",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControlNoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectionReservation"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/selection/reservations/{reservation_id}": {
      "get": {
        "operationId": "selectionGetReservation",
        "security": [
          {
            "SelectionBearer": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ReservationID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControlNoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectionReservation"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      },
      "patch": {
        "operationId": "selectionModifyReservation",
        "security": [
          {
            "SelectionBearer": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ReservationID"
          },
          {
            "$ref": "#/components/parameters/ReservationToken"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/XRequestID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModifyReservationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControlNoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectionReservation"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/selection/reservations/{reservation_id}/release": {
      "post": {
        "operationId": "selectionReleaseReservation",
        "security": [
          {
            "SelectionBearer": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ReservationID"
          },
          {
            "$ref": "#/components/parameters/ReservationToken"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/XRequestID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyObject"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControlNoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseReservationResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AdminBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "PartnerBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "ScannerBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "SelectionBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque, time-bounded BuyerSelectionSession capability."
      }
    },
    "parameters": {
...
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Durable idempotency key scoped to the authenticated actor and operation.",
        "schema": {
          "type": "string",
          "minLength": 1
        }
      },
...
    },
    "responses": {
      "ErrorResponse": {
        "description": "Machine-readable error response.",
        "headers": {
          "X-Request-ID": {
            "$ref": "#/components/headers/XRequestID"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "PartnerEvent": {
        "type": "object",
        "required": ["id", "name", "state", "server_time"],
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "state": {
            "type": "string",
            "enum": ["DRAFT", "ON_SALE", "PAUSED", "SALES_CLOSED", "COMPLETED", "CANCELLED"]
          },
          "starts_at": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "ends_at": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "sales_open_at": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "sales_close_at": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "admission_open_at": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "admission_close_at": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "timezone_name": {
            "type": ["string", "null"]
          },
          "server_time": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "GenericObject": {
        "type": "object",
        "additionalProperties": true
      },
...
      "EmptyObject": {
        "type": "object",
        "additionalProperties": false
      },
...
      "SelectionSession": {
        "type": "object",
        "required": ["id", "event_id", "return_url", "state", "expires_at"],
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^sel_"
          },
          "event_id": {
            "type": "string",
            "pattern": "^evt_"
          },
          "buyer_session_ref": {
            "type": "string"
          },
          "return_url": {
            "type": "string",
            "format": "uri"
          },
          "state": {
            "type": "string",
            "enum": ["ACTIVE", "REVOKED", "EXPIRED"]
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "server_time": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "headers": {
...
      "CacheControlNoStore": {
        "description": "Reservation responses containing protected state must not be cached.",
        "schema": {
          "type": "string",
          "const": "no-store"
        }
      }
    }
  }
}